requirements.txt
fastapi
uvicorn
torch
transformers
motor
pydantic[email]
python-dotenv
passlib
python-jose
pytesseract
Pillow
